# services/engine/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Chatbot Engine API
  description: |
    REST API for the Chatbot Engine — an NLP-powered chatbot platform that supports
    multi-group configuration, intent recognition, query execution, learning feedback,
    and administrative management.
  version: 1.0.0
  contact:
    name: Chatbot Engine Team
servers:
  - url: http://localhost:4000
    description: Local development
  - url: '{productionUrl}'
    description: Production
    variables:
      productionUrl:
        default: https://chatbot-engine.example.com
        description: Production server URL
...
security: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: |
        API key required for all `/api/admin/*` routes when the `ENGINE_API_KEY`
        environment variable is set. Can also be passed as a Bearer token in the
        Authorization header.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Internal server error
...
    ChatRequest:
      type: object
      required:
        - text
        - sessionId
      properties:
        text:
          type: string
          example: How many open tickets are there?
        sessionId:
          type: string
          example: sess_abc123
        groupId:
          type: string
          example: default
        platform:
          type: string
          enum: [web, widget, teams, slack]
          example: web
        feedbackType:
          type: string
          description: Feedback signal (thumbs_up, thumbs_down, rephrase)
          example: thumbs_up
        previousMessageText:
          type: string
          description: The bot message the feedback refers to
        explicitFilters:
          type: object
          additionalProperties:
            type: string
          description: Key-value filter overrides for the query
          example:
            status: open
            region: US
    ChatResponse:
      type: object
      properties:
        text:
          type: string
          example: There are 42 open tickets.
        intent:
          type: string
          example: query.open_tickets
        confidence:
          type: number
          example: 0.93
        richContent:
          type: object
          description: Optional structured data (tables, charts, cards)
        executionMs:
          type: integer
          example: 320

    # ── Health ────────────────────────────────────────────────────────────
...
    QueryStat:
      type: object
      properties:
        id:
          type: string
          example: stat_1710000000_ab12
        queryName:
          type: string
          example: open_tickets
        success:
          type: boolean
        durationMs:
          type: integer
        timestamp:
          type: string
          format: date-time
        filters:
          type: object
          additionalProperties:
            type: string
        error:
          type: string
    AggregatedStat:
      type: object
      properties:
        queryName:
          type: string
        totalExecutions:
          type: integer
        successCount:
          type: integer
        failureCount:
          type: integer
        failureRate:
          type: string
          example: '2.5%'
        avgDurationMs:
          type: integer
        p50Ms:
          type: integer
        p95Ms:
          type: integer

    # ── User Info ─────────────────────────────────────────────────────────
...
    CreateGroupRequest:
      type: object
      required:
        - groupId
        - name
      properties:
        groupId:
          type: string
          pattern: '^[a-z0-9_]+$'
          example: support_team
        name:
          type: string
          example: Support Team
        description:
          type: string
        sources:
          type: array
          items:
            type: string
        apiBaseUrl:
          type: string
        greeting:
          type: string
        helpText:
          type: string
        queries:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              description:
                type: string
              source:
                type: string
              url:
                type: string
              estimated_duration:
                type: integer
              filters:
                type: string
        faq:
          type: array
          items:
            type: object
            properties:
              question:
                type: string
              intent:
                type: string
              answer:
                type: string

    # ── Admin: Query ──────────────────────────────────────────────────────
...
    AdminUser:
      type: object
      properties:
        id:
          type: string
          example: u1
        name:
          type: string
        email:
          type: string
        userid:
          type: string
        brid:
          type: string
        role:
          type: string
          enum: [admin, viewer]
        createdAt:
          type: string
          format: date-time
        updatedBy:
          type: string
        updatedOn:
          type: string
          format: date-time
...
paths:
...
  /api/chat:
    post:
      tags: [Chat]
      summary: Send a message to the chatbot
      description: |
        Processes a user message through the NLP engine and returns the bot response.
        Supports multi-group routing, explicit filter overrides, and feedback signals.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            example:
              text: How many open tickets are there?
              sessionId: sess_abc123
              groupId: default
              platform: web
      responses:
        '200':
          description: Bot response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid message format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Invalid message format
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ════════════════════════════════════════════════════════════════════════
  # QUERIES (public)
  # ════════════════════════════════════════════════════════════════════════
...
  /api/stats:
    get:
      tags: [Stats]
      summary: Get query execution statistics
      parameters:
        - name: queryName
          in: query
          schema:
            type: string
          description: Filter stats to a specific query
        - name: since
          in: query
          schema:
            type: string
            format: date-time
          description: Only include stats after this timestamp
      responses:
        '200':
          description: Statistics with aggregations
          content:
            application/json:
              schema:
                type: object
                properties:
                  stats:
                    type: array
                    items:
                      $ref: '#/components/schemas/QueryStat'
                  aggregated:
                    type: array
                    items:
                      $ref: '#/components/schemas/AggregatedStat'
                  totalRecords:
                    type: integer
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Stats]
      summary: Record a query execution statistic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - queryName
              properties:
                queryName:
                  type: string
                  example: open_tickets
                success:
                  type: boolean
                  default: true
                durationMs:
                  type: integer
                  default: 0
                filters:
                  type: object
                  additionalProperties:
                    type: string
                error:
                  type: string
      responses:
        '201':
          description: Stat recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryStat'
        '400':
          description: Missing queryName
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: queryName is required
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ════════════════════════════════════════════════════════════════════════
  # USER
  # ════════════════════════════════════════════════════════════════════════
...
  /api/admin/groups/create:
    post:
      tags: [Admin/Groups]
      summary: Create a new group via onboarding
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  groupId:
                    type: string
                  queriesAdded:
                    type: integer
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/admin/users/{id}:
    get:
      tags: [Admin/Users]
      summary: Get a user by ID
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags: [Admin/Users]
      summary: Update a user
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                userid:
                  type: string
                brid:
                  type: string
                role:
                  type: string
                  enum: [admin, viewer]
                updatedBy:
                  type: string
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Admin/Users]
      summary: Delete a user
      description: Cannot delete the last admin user.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  deletedUserId:
                    type: string
        '400':
          description: Cannot delete last admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ════════════════════════════════════════════════════════════════════════
  # ADMIN — FILES (Knowledge Base)
  # ════════════════════════════════════════════════════════════════════════
...